ocr-ai-mac/requirements.txt
paddleocr>=2.6.0.3
paddlepaddle>=2.4.0
pika>=1.3.2
python-dotenv>=1.0.0
openai>=1.0.0
requests>=2.31.0
urllib3>=2.0.0
PyMuPDF>=1.23.0